+++ src/components/GlobalFooter/GlobalFooter.tsx
@@ -1,6 +1,6 @@
 export default function GlobalFooter() {
   return (
-    <footer className="mt-20 border-t-2 border-black py-4 text-center">
+    <footer className="border-t-2 border-black py-4 text-center">
       <p>© 2025 LoGravel. All Rights Reserved. </p>
     </footer>
   );