+++ src/components/GlobalHeader/GlobalHeader.tsx
@@ -2,8 +2,8 @@ import Link from 'next/link';
 
 export default function GlobalHeader() {
   return (
-    <header className="sticky top-9 container">
-      <div className="mt-9 flex items-center justify-between rounded-full bg-[#fbfbfb] px-8 py-4 shadow-[0_0_3px_rgba(0,0,0,0.50)]">
+    <header className="sticky top-0 shadow-[0_0_3px_rgba(0,0,0,0.50)]">
+      <div className="container flex items-center justify-between bg-[#fbfbfb] py-4">
         <Link href="/">
           <h1 className="font-space-grotesk text-[32px] font-bold text-[#2D232E]">
             LoGravel